Refs 10180. Loan manager moved to implementation folder.
