+++ src/cpp/fastdds/subscriber/DataReaderImpl/Readme.md
@@ -1,0 +1,7 @@
+## Internal support classes for DataReaderImpl
+
+This folder contains the declaration of several internal support classes to improve
+maintenance and readability of the DataReaderImpl class.
+
+They are not meant to be used outside DataReaderImpl and their functionality should be
+fully covered by the unit tests of the DataReader(Impl).
